requirements.txt
Pillow
streamlit 